Revert "Get all file path resolvers to do things the same way"
